Move Armmorvax processing to the Import helper so it can be called without an HTTP request
